clean up spec helper
