+++ spec/lib/yaml/eo.yml
@@ -1,0 +1,2 @@
+eo:
+  test_string: "test string"
